requirements.txt
chromadb
pypdf
openai
fastapi
jinja2
python-multipart
pymupdf
opencv-python
pillow
requests